+++ app/routes.ts
@@ -9,7 +9,7 @@ export default [
 	index("routes/home.tsx"),
 	route("join", "./routes/auth/join.tsx"),
 	route("login", "./routes/auth/login.tsx"),
-	route("logout", "./routes/auth/logout.tsx"),
+	route("logout", "./routes/auth/logout.ts"),
 
 	route("lorem", "./routes/lorem.tsx"),
 	route("privacy", "./routes/privacy.tsx"),
@@ -19,17 +19,14 @@ export default [
 		index("./routes/userFiles.tsx"),
 		...prefix("files/:fileId", [
 			index("./routes/fileDetail.tsx"),
-			...prefix("download", [
-				index("./routes/fileConfirm.tsx"),
-				route("execute", "./routes/fileDownload.ts"),
-			]),
+			route("download", "./routes/fileDownload.ts"),
 		]),
 	]),
 	route("dashboard", "./routes/user/dashboard.tsx"),
 	...prefix("settings", [route("profile", "./routes/user/profile.tsx")]),
 	...prefix("files", [
 		index("./routes/files.tsx"),
-		route("new", "./routes/user/newFile.tsx"),
+		route("new", "./routes/user/createFile.tsx"),
 		route(":fileId/edit", "./routes/user/editFile.tsx"),
 	]),
 ] satisfies RouteConfig;
